requirements.txt
blinker==1.9.0
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.1.0
fsspec==2024.10.0
huggingface-hub==0.26.5
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.3
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.1
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.1
packaging==24.2
pillow==11.0.0
pip==24.2
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.5
setuptools==75.1.0
sympy==1.13.1
timm==1.0.12
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.11.0
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
gunicorn==23.0.0